models/engine/db_storage.py
#!/usr/bin/python3
<<<<<<< HEAD
"""Defines the DBStorage engine."""
from os import getenv
from models.base_model import Base
from models.base_model import BaseModel
from models.amenity import Amenity
from models.city import City
from models.place import Place
from models.review import Review
from models.state import State
from models.user import User
from sqlalchemy import create_engine
from sqlalchemy.orm import relationship
from sqlalchemy.orm import scoped_session
from sqlalchemy.orm import sessionmaker


class DBStorage:
    """Represents a database storage engine.

    Attributes:
        __engine (sqlalchemy.Engine): The working SQLAlchemy engine.
        __session (sqlalchemy.Session): The working SQLAlchemy session.
    """
=======
"""This module defines a class that manages database storage for hbnb clone"""
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker, scoped_session
import os

from models.base_model import BaseModel, Base
from models.user import User
from models.place import Place
from models.amenity import Amenity
from models.state import State
from models.review import Review
from models.city import City

HBNB_ENV = os.environ.get("HBNB_ENV")
HBNB_MYSQL_USER = os.environ.get("HBNB_MYSQL_USER")
HBNB_MYSQL_PWD = os.environ.get("HBNB_MYSQL_PWD")
HBNB_MYSQL_HOST = os.environ.get("HBNB_MYSQL_HOST")
HBNB_MYSQL_DB = os.environ.get("HBNB_MYSQL_DB")


class DBStorage(object):
    """Database storage"""
>>>>>>> 1512bbc73944390258cdb9d8055800254a97323e

    __engine = None
    __session = None

<<<<<<< HEAD
    def __init__(self):
        """Initialize a new DBStorage instance."""
        self.__engine = create_engine("mysql+mysqldb://{}:{}@{}/{}".
                                      format(getenv("HBNB_MYSQL_USER"),
                                             getenv("HBNB_MYSQL_PWD"),
                                             getenv("HBNB_MYSQL_HOST"),
                                             getenv("HBNB_MYSQL_DB")),
                                      pool_pre_ping=True)
        if getenv("HBNB_ENV") == "test":
            Base.metadata.drop_all(self.__engine)

    def all(self, cls=None):
        """Query on the curret database session all objects of the given class.

        If cls is None, queries all types of objects.

        Return:
            Dict of queried classes in the format <class name>.<obj id> = obj.
        """
        if cls is None:
            objs = self.__session.query(State).all()
            objs.extend(self.__session.query(City).all())
            objs.extend(self.__session.query(User).all())
            objs.extend(self.__session.query(Place).all())
            objs.extend(self.__session.query(Review).all())
            objs.extend(self.__session.query(Amenity).all())
        else:
            if type(cls) == str:
                cls = eval(cls)
            objs = self.__session.query(cls)
        return {"{}.{}".format(type(o).__name__, o.id): o for o in objs}

    def new(self, obj):
        """Add obj to the current database session."""
        self.__session.add(obj)

    def save(self):
        """Commit all changes to the current database session."""
        self.__session.commit()

    def delete(self, obj=None):
        """Delete obj from the current database session."""
        if obj is not None:
            self.__session.delete(obj)

    def reload(self):
        """Create all tables in the database and initialize a new session."""
        Base.metadata.create_all(self.__engine)
        session_factory = sessionmaker(bind=self.__engine,
                                       expire_on_commit=False)
        Session = scoped_session(session_factory)
        self.__session = Session()

    def close(self):
        """Close the working SQLAlchemy session."""
        self.__session.close()
=======
    classes = {
        "State": State,
        "City": City,
        "User": User,
        "Place": Place,
        "Amenity": Amenity,
        "Reviewer": Review
    }

    def __init__(self):
        """Instantiate a DBStorage instance"""

        self.__engine = create_engine(
            "mysql+mysqldb://{}:{}@{}/{}".format(
                HBNB_MYSQL_USER, HBNB_MYSQL_PWD, HBNB_MYSQL_HOST, HBNB_MYSQL_DB
            ),
            pool_pre_ping=True,
        )

        if HBNB_ENV == "test":
            Base.metadata.drop_all(self.__engine)

    def all(self, cls=None):
        """Return all objects on current database session of class cls.
        If cls is None return all objects.

        Args:
            cls (class): Class of the objects to return. Defaults to None.
        """

        if cls:
            objs = self.__session.query(cls).all()
        else:
            objs = []
            for key in self.classes.keys():
                obj = self.__session.query(self.classes[key]).all()
                objs.extend(obj)

        objs_dict = {}
        for obj in objs:
            key = f"{obj.__class__.__name__}.{obj.id}"
            objs_dict[key] = obj
        return objs_dict

    def new(self, obj):
        """Add obj to the current database session

        Args:
            obj (object): Object to add to the session
        """

        self.__session.add(obj)

    def save(self):
        """Commit all changes of the current database session"""

        self.__session.commit()

    def delete(self, obj=None):
        """Deletes obj from the current database session if not node.

        Args:
            obj (object): Object to delete. Defaults to None.
        """

        if obj:
            self.__session.delete(obj)

    def reload(self):
        """Create all tables in the database"""

        Base.metadata.create_all(self.__engine)

        session_factory = sessionmaker(
            bind=self.__engine,
            expire_on_commit=False
        )
        Session = scoped_session(session_factory)
        self.__session = Session()
>>>>>>> 1512bbc73944390258cdb9d8055800254a97323e
